Improve documentation and report of time measures
Make explicit that all reported times in mt_measures (except idle_time)
are reported in timestamp metric. Ensure that in special case where
first timestamp is > 0, times are still reported in timestamp metric (so
far, a period without movement starting at timestamp 0 was assumed and a
corresponding timestamp was added).