Fix bug in mt_exclude_initiation (resolves #14)
This fixes a bug in mt_exclude_initiation so that the correct initial position is used when removing the initial period without movement in a trial. Previously, mt_exclude_initiation was only behaving as intended when the initial position was (0,0), which probably was the case in many settings (in particular, if the trajectories were aligned using mt_align_start with default start values). If the initial position in a trial corresponded to a different value, the initial period without movement in most cases was simply not removed. This has now been fixed.