Replace only_ggplot with return_type in mt_plot
Replace only_ggplot with return_type argument in mt_plot and mt_plot_aggregate functions. The new return_type argument allows for further customization of the plots: If return_type == "plot" (the default), a new ggplot is created and the trajectories are plotted using geom_path (corresponds to the old behavior where only_ggplot was set to FALSE, the default). If return_type == "mapping", only the ggplot object containing the mapping but without any geoms is returned (corresponds to the old behavior where only_ggplot was set to TRUE). If return_type == "geoms", only the geoms are returned, which allows adding the plotted trajectories to an existing ggplot (new functionality). Examples were adapted accordingly.